updated csv functions

+++ app/database/migrations/2014_09_08_061222_create_vocab_scores_table.php
@@ -33,7 +33,9 @@ class CreateVocabScoresTable extends Migration {
 	 */
 	public function down()
 	{
-        Schema::drop('vocab_scores');
+        if (Schema::hasTable('vocab_scores')) {
+            Schema::drop('vocab_scores');
+        }
 	}
 
 }
